+++ platform/60D.109/consts.h
@@ -110,7 +110,7 @@
  */
 #define AJ_LCD_Palette 0x3BFA4
 
-#define LV_BOTTOM_BAR_DISPLAYED (((*(int*)0x5680) == 0xF) || ((*(int*)0x2A434) != 0x17)) // dec CancelBottomInfoDispTimer
+#define LV_BOTTOM_BAR_DISPLAYED (((*(int8_t*)0x5680) == 0xF) || ((*(int8_t*)0x2A434) != 0x17)) // dec CancelBottomInfoDispTimer
 #define ISO_ADJUSTMENT_ACTIVE ((*(int*)0x5680) == 0xF) // dec ptpNotifyOlcInfoChanged
 
 #define COLOR_FG_NONLV 1
